+++ src/main/java/com/nodemules/rhp/repository/VenueRepository.java
@@ -7,5 +7,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
  * @author brent
  * @since 7/29/17.
  */
-public interface VenueRepository extends JpaRepository<Venue, Integer> {
+public interface VenueRepository extends JpaRepository<Venue, Long> {
 }
